ㅁㄴㅇㄻㄴㅇㄹ

+++ src/threads/thread.c
@@ -224,10 +224,10 @@ void compare_wait_left_tick(const struct list_elem *a, const struct list_elem *b
   struct thread * b_thread = list_entry(b,struct thread,blockelem);
   int64_t a_leftTick = get_remain_time(a_thread);
   int64_t b_leftTick = get_remain_time(b_thread);
-  return a_leftTick<b_leftTick;
+  return a_leftTick < b_leftTick;
 }
 int get_remain_time(struct thread * thread){
-  return (int)thread->wait_ticks - timer_elapsed(thread->wait_start_time);
+  return (thread->wait_ticks) - timer_elapsed(thread->wait_start_time);
 }
 void check_wait_threads(){
   struct list_elem * a0;
